Work on setConfigItem etc.

+++ Emulator/Agnus/Blitter/BlitterTypes.h
@@ -17,7 +17,7 @@
 
 typedef struct
 {
-    int accuracy;
+    isize accuracy;
 }
 BlitterConfig;
